Include "start" in the try/finally in docs
Both producer and consumer classes have a "start" and "stop" method.

`start` initializes background co-routine, open connections and execute
various tasks that need to be cleaned up with the `stop` method.

But `stop` also hold a logic based on a boolean `_closed` that is
initialized to `False` at the end of the `__init__` method.

Thus, if one create a consumer or a producer and doesn't call `stop`
even if `start` was never call, the finalizer in `__del__` will send a
warning.

A correct pattern of calling `stop` in a try/finally block seems to be
then to do the `try` just after the `__init__` and not just after the
`start` as an error or a task cancellation in between would leave the
producer/consumer "open" and trigger the warning

fixes #938