Resolve API versions on connection
As stated in the Kafka protocol documentation https://kafka.apache.org/protocol#api_versions
the API versions should be solved for each single connection we open
with a broker (the broker might be updated while we talk to it).

This commit aim to move all the versions logic into the protocol and the
connection only instead of being spread in various classes.

It is introducing for that a new "Request" class that is acting as a
builder for the Struct we are going to send into the wire. The Request
class tries to collect all possible parameters, and then will react
according the best available version in two possible ways:

* failing, as the intent of the caller was clearly to use a new feature
  that is not available on the broker yet
* best effort, as the caller can totally work with an older version of
  the API

This choice is made on per API basis.

From the caller perspective, if a different logic need to be ran
according the API version, it can look up on the Response.API_VERSION

Doing such a change will help aiokafka supporting Kafka 4.0 as some
API versions were removed as mentioned in #1085

It might not achieve the compatibility as we could lack some protocol
definition, but it will definitively help.