Simplify flexible versions
The flexible versions is a protocol specificity for newer versions of
the API. When an API is flexible, it is using more compact structures
and also allow additional "dynamic" fields that could be added without
the need to introduce a new API versions.

This commit move the flexible versions support to the protocol layer, so
it is more transparent and easy when defining Struct classes and
schemas.

When defining the schema, we can specify a tagged field with a tuple
containing the field name and the field tag.