Handle node failure properly in sender (#1135)
Various sender activities require the sender to talk to a dedicated
coordinator node, for a related group or for transaction management.

When trying to reach this node, an error could happen because of the
broker being unavailable or unreachable. In this case, we must mark the
coordinator dead, so instead of retrying over and over to the same node,
we will first try to find an active coordinator.

fixes #1134

Co-authored-by: Vincent Maurin <vincent.maurin@checkstep.com>