Use linger_ms to delay batches
Currently, the linger_ms parameter is used to limit the number of
request sent to a given broker as adding at minimum linger_ms between to
produce requests. It doesn't produce the expected result when the max
size of batches is also reached.

The purpose of linger_ms should be not limiting the rate of request, but
instead keeping a batch still open long enough so it has the opportunity
to grow bigger.

Moving linger_ms to the message accumulator (like in the java client)
gives the opportunity to have a condition both on the max size and the
lingering time.

fixes #1137